Address PET review summary follow-ups
